libavcodec: v4l2m2m: include SPS/PPS headers in keyframes
This adds an option "-repeat_seq_header", which tells the encoder to
insert sequence headers repeatedly in the stream. This is required for
clients to be able to join mid-stream in a H.264 stream from the
Raspberry Pi.

Without this option (the previous and default behavior), PPS/SPS
headers are only added to the first IDR frame. When
-repeat_seq_header is enabled, PPS/SPS is added to every IDR frame,
allowing clients to decode the stream from that point.

Signed-off-by: Marcus Better <marcus@better.se>